requirements.txt
mypy==1.5.1
Django==4.2.6
djangorestframework==3.14.0
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
alembic==1.7.3
aiofiles==0.8.0
Jinja2==3.0.2
